IPEHR-160 refacor query rows code
